+++ frontend/src/components/fillers/OrderItemListView.tsx
@@ -5,22 +5,11 @@ import { useNavigate, useParams } from 'react-router-dom';
 
 function OrderItemListView() {
     const { selectedOrder } = useSelectedOrder();
-    const navigate = useNavigate();
-    const { id } = useParams();
-
-    const startFilling = () => {
-        navigate(`/dashboard/filler/order/${id}`);
-    };
 
     if (!selectedOrder) return null;
 
     return (
         <>
-            {selectedOrder?.orderStatus === OrderStatus.ACCEPTED && (
-                <Group justify="center" my="md">
-                    <Button onClick={startFilling}>Begin Filling</Button>
-                </Group>
-            )}
             <Paper shadow="xs" p="md">
                 {selectedOrder?.items?.map((item) => (<>
                     <Group key={item.id} justify={'space-between'}>